Updated Battery Percentage to be more accurate
Changed the way that percentage is calculated. Instead of using the design capacity for the calculation now the last full capacity measured is used. This is helpful as I use a thinkpad from 2011 and the battery is only at 80% capacity. As a result when the battery was fully charged it would only show 80% because it is 80% of the designed capacity. Now when the battery is full it will show 100%